backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
requests
google-generativeai
sentence-transformers
numpy
